requirements.txt
fastapi>=0.120.0
uvicorn>=0.24.0
streamlit>=1.28.0
speechrecognition>=3.10.0
google-generativeai>=0.8.0
python-multipart>=0.0.6
websockets>=12.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
noisereduce>=2.0.0
webrtcvad>=2.0.10
pydub>=0.25.0
faster-whisper
pyannote.audio==4.0.1